requirements.txt
certifi==2018.4.16
chardet==3.0.4
Django==2.0.6
djangorestframework==3.8.2
idna==2.6
numpy==1.14.5
psycopg2==2.6.2
Pygments==2.2.0
pytz==2018.4
requests==2.18.4
scipy==1.1.0
urllib3==1.22
